M    libOPC_UA/libOPC_UA.cpp
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2126 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/OPC_UA/libOPC_UA/libOPC_UA.cpp
@@ -2096,7 +2096,7 @@ nextReq:
     try {
 	if(rba.size() < 8) return false;
 	mSz = iNu(rba, off, 4);
-	rb = rba.substr(0, std::min(mSz,rba.size()));
+	rb = rba.substr(0, std::min(mSz,(uint32_t)rba.size()));
 
 	//Check for hello message type
 	if(rb.compare(0,4,"HELF") == 0) {
